refactor(InputValidatorTest): 분리되어있던 클래스 병합
